requirements.txt
flask==2.0.1
werkzeug==2.0.3
python-dotenv==0.19.0
openai>=1.0.0
gunicorn==20.1.0